+++ src/main/java/com/nhnacademy/store99/bookstore/book_author/service/BookAuthorServiceImpl.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.store99.bookstore.book_author.service;
 
 import com.nhnacademy.store99.bookstore.book.response.BookResponse;
+import com.nhnacademy.store99.bookstore.book.response.BookTransDTO;
 import com.nhnacademy.store99.bookstore.book_author.repository.impl.BookAuthorRepositoryImpl;
-import com.nhnacademy.store99.bookstore.book_author.response.BookTransDTO;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
