+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/response/CategoryParentsDTO.java
@@ -9,7 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CategoryParentsDTO {
     Long categoryId;
-    String categoryName;
-    int categoryDepth;
     Long parentCategoryId;
 }
