+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/service/BookCategoryService.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book_category.service;
 
-import com.nhnacademy.store99.bookstore.book_author.response.BookTransDTO;
+import com.nhnacademy.store99.bookstore.book.response.BookTransDTO;
 import com.nhnacademy.store99.bookstore.book_category.response.CategoryParentsDTO;
 import java.util.List;
 import org.springframework.data.domain.Page;
@@ -9,5 +9,5 @@ import org.springframework.data.domain.Pageable;
 public interface BookCategoryService {
     List<CategoryParentsDTO> getBooksByCategory(Long categoryId);
 
-    Page<BookTransDTO> getBooksByCategories(List<CategoryParentsDTO> parentsDTOList, Pageable pageable);
+    Page<BookTransDTO> getBooksByCategories(Long categoryId, Pageable pageable);
 }
