+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/service/Impl/BookCategoryServiceImpl.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.store99.bookstore.book_category.service.Impl;
 
-import com.nhnacademy.store99.bookstore.book_author.response.BookTransDTO;
+import com.nhnacademy.store99.bookstore.book.response.BookTransDTO;
 import com.nhnacademy.store99.bookstore.book_category.repository.BookCategoryRepository;
 import com.nhnacademy.store99.bookstore.book_category.response.CategoryParentsDTO;
 import com.nhnacademy.store99.bookstore.book_category.service.BookCategoryService;
@@ -21,7 +21,8 @@ public class BookCategoryServiceImpl implements BookCategoryService {
     }
 
     @Override
-    public Page<BookTransDTO> getBooksByCategories(List<CategoryParentsDTO> parentsDTOList, Pageable pageable) {
-        return bookCategoryRepository.getBooksByCategories(parentsDTOList, pageable);
+    public Page<BookTransDTO> getBooksByCategories(Long categoryId, Pageable pageable) {
+        List<CategoryParentsDTO> cp = bookCategoryRepository.getCategoriesByParentsId(categoryId);
+        return bookCategoryRepository.getBooksByCategories(cp, pageable);
     }
 }
